Refactor av telefonnummer
