+++ src/pages/endre-personopplysninger/EndreOpplysningerView.tsx
@@ -3,7 +3,7 @@ import AdresserVisning from "pages/forside/sections/4-personinfo/3-adresser/Adre
 import Utbetalinger from "pages/forside/sections/4-personinfo/4-utbetalinger/Utbetalinger";
 import Box from "components/box/Box";
 import kontaktIkon from "assets/img/Kontakt.svg";
-import TelefonnummerHosNav from "pages/forside/sections/4-personinfo/2-kontaktinfo/subsections/TelefonnummerHosNav";
+import TelefonnummerHosNav from "pages/forside/sections/4-personinfo/2-kontaktinfo/subsections/telefonnummer/TelefonnummerHosNav";
 import { Personalia } from "../../types/personalia";
 import { Adresser } from "../../types/adresser";
